Associate users with todos

+++ server/server.js
@@ -17,9 +17,10 @@ const port = process.env.PORT || 3000;  //  defind port for heroku
 app.use(bodyParser.json());
 //app.use(morgan('dev'));
 
-app.post('/todos', (req,res) => {
+app.post('/todos', authenticate, (req,res) => {
     var todo = new Todo({
-        text: req.body.text
+        text: req.body.text,
+        _creator: req.user._id
     });
 
     todo.save().then((doc) => {
@@ -29,21 +30,21 @@ app.post('/todos', (req,res) => {
     })
 });
 
-app.get('/todos', (req,res) => {
-    Todo.find()
-        .then((todos) => {
-            res.send({todos})
-        })
-        .catch((e) => res.status(400).send);
+app.get('/todos', authenticate, (req,res) => {
+    Todo.find({_creator:req.user._id})
+    .then((todos) => {
+         res.send({todos})
+    })
+    .catch((e) => res.status(400).send);
 });
 
-app.get('/todos/:id', (req, res) => {
+app.get('/todos/:id', authenticate, (req, res) => {
     var id = req.params.id;
 
     if(!ObjectID.isValid(id))
         return res.status(404).send();
 
-    Todo.findById(id)
+    Todo.findOne({_id: id, _creator:req.user._id})
         .then((todo) => {
             if(todo)
                 res.status(200).send({todo});
@@ -64,13 +65,13 @@ app.get('/todos/:id', (req, res) => {
             //  400 - and send empty body back
 });
 
-app.delete('/todos/:id', (req, res) => {
+app.delete('/todos/:id', authenticate, (req, res) => {
     var id = req.params.id;
 
     if(!ObjectID.isValid(id))
         return res.status(404).send();
 
-    Todo.findOneAndDelete({_id : id})
+    Todo.findOneAndDelete({_id : id, _creator:req.user._id})
     .then((todo) =>{
         if(!todo)
             return res.status(404).send();
@@ -82,7 +83,7 @@ app.delete('/todos/:id', (req, res) => {
     
 });
 
-app.patch('/todos/:id', (req,res) => {
+app.patch('/todos/:id', authenticate, (req,res) => {
     var id = req.params.id;
     var body = _.pick(req.body, ['text', 'completed']);
 
@@ -97,7 +98,7 @@ app.patch('/todos/:id', (req,res) => {
         body.completedAt = null;
     }
 
-    Todo.findByIdAndUpdate(id, {$set: body}, {new:true})
+    Todo.findOneAndUpdate({_id:id, _creator:req.user._id}, {$set: body}, {new:true})
     .then((todo) => {
         if(!todo)
             return res.status(404).send();
